+++ src/test/java/com/structurizr/dsl/GroupParserTests.java
@@ -9,9 +9,9 @@ class GroupParserTests extends AbstractTests {
     private GroupParser parser = new GroupParser();
 
     @Test
-    void test_parse_ThrowsAnException_WhenThereAreTooManyTokens() {
+    void parse_ThrowsAnException_WhenThereAreTooManyTokens() {
         try {
-            parser.parse(tokens("group", "name", "extra"));
+            parser.parse(null, tokens("group", "name", "extra"));
             fail();
         } catch (Exception e) {
             assertEquals("Too many tokens, expected: group <name> {", e.getMessage());
@@ -19,9 +19,9 @@ class GroupParserTests extends AbstractTests {
     }
 
     @Test
-    void test_parse_ThrowsAnException_WhenTheNameIsMissing() {
+    void parse_ThrowsAnException_WhenTheNameIsMissing() {
         try {
-            parser.parse(tokens("group"));
+            parser.parse(null, tokens("group"));
             fail();
         } catch (Exception e) {
             assertEquals("Expected: group <name> {", e.getMessage());
@@ -29,9 +29,33 @@ class GroupParserTests extends AbstractTests {
     }
 
     @Test
-    void test_parse() {
-        ElementGroup group = parser.parse(tokens("group", "Group 1"));
+    void parse() {
+        ElementGroup group = parser.parse(context(), tokens("group", "Group 1"));
         assertEquals("Group 1", group.getName());
+        assertTrue(group.getElements().isEmpty());
+    }
+
+    @Test
+    void parse_NestedGroup_ThrowsAnExceptionWhenNestedGroupsAreNotConfigured() {
+        ModelDslContext context = new ModelDslContext(new ElementGroup(workspace.getModel(), "Group 1"));
+        context.setWorkspace(workspace);
+
+        try {
+            parser.parse(context, tokens("group", "Group 2"));
+            fail();
+        } catch (Exception e) {
+            assertEquals("To use nested groups, please define a model property named structurizr.group.separator", e.getMessage());
+        }
+    }
+
+    @Test
+    void parse_NestedGroup() {
+        workspace.getModel().addProperty("structurizr.group.separator", "/");
+        ModelDslContext context = new ModelDslContext(new ElementGroup(workspace.getModel(), "Group 1"));
+        context.setWorkspace(workspace);
+
+        ElementGroup group = parser.parse(context, tokens("group", "Group 2"));
+        assertEquals("Group 1/Group 2", group.getName());
         assertTrue(group.getElements().isEmpty());
     }