+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.29.0 (unreleased)
+
+- Adds support for nested groups.
+- Upgrades structurizr/java to [v1.23.0](https://github.com/structurizr/java/releases/tag/v1.23.0).
+
 ## 1.28.3 (11th March 2023)
 
 - Upgrades structurizr/java to [v1.22.3](https://github.com/structurizr/java/releases/tag/v1.22.3).